work on rx tia cal

+++ include/LMS7002M/LMS7002M.h
@@ -632,10 +632,9 @@ LMS7002M_API void LMS7002M_rxtsp_tsg_tone(LMS7002M_t *self, const LMS7002M_chan_
  * Read the digital RSSI indicator in the Rx TSP chain.
  * \param self an instance of the LMS7002M driver
  * \param channel the channel LMS_CHA or LMS_CHB
- * \return the RSSI as a fraction of the full scale (max 1.0)
+ * \return the RSSI as an unsigned 16-bit number
  */
-LMS7002M_API double LMS7002M_rxtsp_read_rssi(LMS7002M_t *self, const LMS7002M_chan_t channel);
-
+LMS7002M_API uint16_t LMS7002M_rxtsp_read_rssi(LMS7002M_t *self, const LMS7002M_chan_t channel);
 
 /*!
  * DC offset correction value for Rx TSP chain.
